created a whiteHeader component
