+++ src/screen/feedback/style.css
@@ -8,7 +8,7 @@
   padding-bottom: 0.4rem;
 }
 
-.repPersonalDetails__sectioni {
+.repPersonalDetails__section {
   min-height: 120vh;
 }
