+++ src/screen/login/index.jsx
@@ -1,0 +1,6 @@
+import './style.css'
+
+const Login = () => {
+  return <section></section>
+}
+export default Login
